feat: update loop transition story, remove old stories
